fix(web): address code review feedback for offline mode PR
- Extract shared OfflineMutationResult type to eliminate duplication
- Simplify OfflineIndicator dismiss behavior to avoid React lint issues
- Fix addToExchange test payload to use convocationId (matching action type)
- Add .catch() handler for sync() promise in PendingActionsIndicator
